inject: move out of internal
