+++ handler.go
@@ -8,7 +8,7 @@ import (
 	"net/http"
 	"reflect"
 
-	"github.com/flamego/flamego/internal/inject"
+	"github.com/flamego/flamego/inject"
 )
 
 // Handler is any callable function. Flamego attempts to inject services into
